# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Furtalk 评论系统的 HTTP API：覆盖站点与管理员引导、\n第一方与 widget 认证/会话、评论发布与审核、站点/用户/设置/\n提供商管理与邮件通知退订。",
        "title": "Furtalk API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/providers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-providers"
                ],
                "summary": "列出提供商配置",
                "responses": {
                    "200": {
                        "description": "提供商列表（不含机密；CAPTCHA 项无 enabled）",
                        "schema": {
                            "$ref": "#/definitions/handler.ProvidersResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/providers/{provider_key}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin-providers"
                ],
                "summary": "新增或更新提供商",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提供商 key",
                        "name": "provider_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "提供商配置（CAPTCHA 不允许携带 enabled）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProviderUpsertRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "保存完成"
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "提供商密钥数据损坏",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-providers"
                ],
                "summary": "删除提供商配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提供商 key",
                        "name": "provider_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除完成"
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "提供商不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/providers/{provider_key}/test": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-providers"
                ],
                "summary": "测试提供商连通性",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提供商 key",
                        "name": "provider_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status=ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "提供商不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "服务暂不可用",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-settings"
                ],
                "summary": "读取系统设置",
                "responses": {
                    "200": {
                        "description": "公开设置项列表",
                        "schema": {
                            "$ref": "#/definitions/handler.SettingsResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-settings"
                ],
                "summary": "更新系统设置",
                "parameters": [
                    {
                        "description": "需要修改的设置项",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SettingsPatchRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的完整设置列表",
                        "schema": {
                            "$ref": "#/definitions/handler.SettingsResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/smtp/test": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smtp"
                ],
                "summary": "测试 SMTP 投递配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status=ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "投递失败",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ProviderMetadata": {
            "type": "object",
            "properties": {
                "configured": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "provider_key": {
                    "type": "string"
                },
                "public_config": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "handler.ProviderUpsertRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "enabled": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "handler.ProvidersResponse": {
            "type": "object",
            "properties": {
                "providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ProviderMetadata"
                    }
                }
            }
        },
...
        "handler.SettingsPatchRequest": {
            "type": "object",
            "properties": {
                "settings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/setting.SettingItem"
                    }
                }
            }
        },
        "handler.SettingsResponse": {
            "type": "object",
            "properties": {
                "settings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/setting.SettingItem"
                    }
                }
            }
        },
...
        "httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpx.ErrorBody"
                }
            }
        },
...
        "setting.SettingItem": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/setting.SettingType"
                },
                "value": {}
            }
        },
        "setting.SettingType": {
            "type": "string",
            "enum": [
                "string",
                "integer",
                "boolean",
                "json"
            ],
            "x-enum-varnames": [
                "SettingTypeString",
                "SettingTypeInteger",
                "SettingTypeBoolean",
                "SettingTypeJSON"
            ]
        }
    }
}